adding in ability to use a config file and specify gitea to keybase username mappings so that some notifications can be DM'd rather than blasted on a channel
