+++ .gitignore
@@ -17,3 +17,5 @@ build/
 
 # IDE
 .idea/
+
+keybase-gitea-bot
