+++ example.cfg
@@ -1,0 +1,24 @@
+[BotConfig]
+HTTPPrefix=`http://hostname:8080`
+WebhookSecret=`mysupersecretwebhook`
+GiteaURL=`http://gitea.example.com:8080`
+DirectMessages=true #send DMs to users on assignments/PR/etc
+#KeybaseLocation=/path/to/keybase
+#Home directory for keybase service
+#Home=/path/to/keybase/home
+#Conversation name or ID to announce when the bot begins
+Announcement="abcdef1234"
+#Conversation name or ID to report bot errors to
+ErrReportConv="abcdef1234"
+#Database Source Name
+DSN=`user:password@tcp(dbip)/dbname`
+#MultiDSN="multidbstring"
+#StathatEZKey="examplekey"
+#Allow the bot to read it's own messages (default: false)
+#ReadSelf=false
+
+[UserMaps]
+giteauserA=keybaseuserAA
+giteaUserB=KeyBaseUserBB
+UserOnGiteaC = KeybaseUserCC
+
